feature: WIP OpenSeadragon use IIIF configuration
We use default information from a web page. It requires setting up width and height values. We use some assumed value, however, those are not correct and part of the image is not pulled with required zoom quality on the first load.

We have code repetition. We should DRY this into a single shared partial. Note this code uses direct script to embeded OpenSeedragon. Some other parts are using openseadragon-rails gem with openseadragon_picture_tag.