feature: Generate and save IIIIF manifest to disk
Add new rake taks that traverese across all Registers, generates its manifests and saves them to local tmp folder. We upload those files to NFS and S3 manually. This is usfull for experimenting and improving Registers manifests fiels.

Rename old task name  for saving manifests to Fedora. Know bug here, we cannot save to Fedora due to predicate URI errors. This is mostl likly related with prevoius migration. It is possible that Fedora 4 is not configured with required predicates, or that ActiveFedora API change and now our model definition are not good for migrated data.